LPS-116693 Stop using URLs to retrieve JSModule content: use InputStream instead
Because the URL class is final, there's no way to override module contents
other than writing a custom URL handler, which is way overkill.

In this commit we change URLs to InputStreams to be able to do it and
deprecate the old URL based methods.
